+++ lib/schemas/conversation_schema_group.dart
@@ -1,18 +1,18 @@
-Map<String, dynamic> conversationSchema({
+Map<String, dynamic> conversationSchemaGroup({
   required String name,
   required String description,
-  required String userUuid,
+  required List<Map<String, dynamic>> userUuid,
   required String notes,
   required int messagesPerBox,
   required bool isGroup,
-  required String conversationUUID
+  required String conversationUUID,
 }) {
   DateTime now = DateTime.now();
   return {
     'schema-version': '1.0.0',
     'name': name,
     'description': description,
-    'userUuid': userUuid,
+    'users': userUuid,
     'createdAt': now,
     'updatedAt': now,
     'notes': notes,