+++ CHANGELOG.md
@@ -1,3 +1,15 @@
 ## 0.0.1
 
 * Initial release.
+
+## 0.0.2
+
+* Add conversation schemas.
+
+* Add message schemas.
+
+* Add create new conversation function.
+
+* Add delete conversation function.
+
+* Improved error handling and clarity.